chore: renaming `src/cmds` to `src/commands`
