+++ __tests__/commands/docs/multiple.test.ts
@@ -5,7 +5,7 @@ import frontMatter from 'gray-matter';
 import nock from 'nock';
 import { describe, beforeAll, beforeEach, afterAll, it, expect } from 'vitest';
 
-import Command from '../../../src/cmds/docs/index.js';
+import Command from '../../../src/commands/docs/index.js';
 import getAPIMock, { getAPIMockWithVersionHeader } from '../../helpers/get-api-mock.js';
 import hashFileContents from '../../helpers/hash-file-contents.js';
 import { runCommand } from '../../helpers/setup-oclif-config.js';
