+++ __tests__/commands/login.test.ts
@@ -2,7 +2,7 @@ import nock from 'nock';
 import prompts from 'prompts';
 import { describe, beforeAll, beforeEach, afterAll, afterEach, it, expect } from 'vitest';
 
-import Command from '../../src/cmds/login.js';
+import Command from '../../src/commands/login.js';
 import APIError from '../../src/lib/apiError.js';
 import configStore from '../../src/lib/configstore.js';
 import getAPIMock from '../helpers/get-api-mock.js';
