+++ src/commands/categories/index.ts
