+++ src/commands/docs/prune.ts
