+++ src/commands/logout.ts
