+++ src/commands/openapi/reduce.ts
