+++ src/commands/openapi/validate.ts
