+++ src/commands/whoami.ts
