+++ src/index.ts
@@ -1,22 +1,22 @@
-import CategoriesCreateCommand from './cmds/categories/create.js';
-import CategoriesCommand from './cmds/categories/index.js';
-import ChangelogsCommand from './cmds/changelogs.js';
-import CustomPagesCommand from './cmds/custompages.js';
-import DocsCommand from './cmds/docs/index.js';
-import DocsPruneCommand from './cmds/docs/prune.js';
-import LoginCommand from './cmds/login.js';
-import LogoutCommand from './cmds/logout.js';
-import OpenCommand from './cmds/open.js';
-import OpenAPIConvertCommand from './cmds/openapi/convert.js';
-import OpenAPICommand from './cmds/openapi/index.js';
-import OpenAPIInspectCommand from './cmds/openapi/inspect.js';
-import OpenAPIReduceCommand from './cmds/openapi/reduce.js';
-import OpenAPIValidateCommand from './cmds/openapi/validate.js';
-import CreateVersionCommand from './cmds/versions/create.js';
-import DeleteVersionCommand from './cmds/versions/delete.js';
-import VersionsCommand from './cmds/versions/index.js';
-import UpdateVersionCommand from './cmds/versions/update.js';
-import WhoAmICommand from './cmds/whoami.js';
+import CategoriesCreateCommand from './commands/categories/create.js';
+import CategoriesCommand from './commands/categories/index.js';
+import ChangelogsCommand from './commands/changelogs.js';
+import CustomPagesCommand from './commands/custompages.js';
+import DocsCommand from './commands/docs/index.js';
+import DocsPruneCommand from './commands/docs/prune.js';
+import LoginCommand from './commands/login.js';
+import LogoutCommand from './commands/logout.js';
+import OpenCommand from './commands/open.js';
+import OpenAPIConvertCommand from './commands/openapi/convert.js';
+import OpenAPICommand from './commands/openapi/index.js';
+import OpenAPIInspectCommand from './commands/openapi/inspect.js';
+import OpenAPIReduceCommand from './commands/openapi/reduce.js';
+import OpenAPIValidateCommand from './commands/openapi/validate.js';
+import CreateVersionCommand from './commands/versions/create.js';
+import DeleteVersionCommand from './commands/versions/delete.js';
+import VersionsCommand from './commands/versions/index.js';
+import UpdateVersionCommand from './commands/versions/update.js';
+import WhoAmICommand from './commands/whoami.js';
 
 export { default as createGHA } from './lib/hooks/createGHA.js';
 export { default as prerun } from './lib/hooks/prerun.js';
